Temporarily introduce a delay after installing Brakeman
We're seeing an intermittent problem where Brakeman sometimes seemingly
fails to install or isn't fully installed by the time we try and run it.

We don't really have any ideas about how to fix it so we're going to
temporarily introduce this delay. I'll give it a month and decide then
whether it's had an effect or not.